CPM-636: Add ExceptionFormatter
