--- src/Akeneo/Pim/Enrichment/Bundle/Controller/ExternalApi/ProductUpdater.php
@@ -1,8 +1,0 @@
-<?php
-
-namespace Akeneo\Pim\Enrichment\Bundle\Controller\ExternalApi;
-
-interface ProductUpdater
-{
-    public function update(array $data): void;
-}
